python_scripts/requirements.txt
pandas>=1.0.0
numpy>=1.18.0
scikit-learn==1.5.1
supabase==2.5.1
python-dotenv==1.0.1
nba-api==1.1.11
pandas==2.0.3
numpy==1.24.4